+++ tools/run_all_checks.php
@@ -33,6 +33,7 @@ $checks = [
     'check_seed_test_data.php',
     'check_fgetcsv_escape.php',
     'check_menu_overlay.php',
+    'check_hamburger_menu_personal.php',
     'check_confirm_modal.php',
     'check_js_syntax.php',
     'check_password_toggle.php',